feat(sales): 报价单/询价单明细优化
- 报价单新增保存后不关弹窗，返回新ID可续改；复制单走新增分支
- 样品明细：编码生成器下沉明细行弹窗、同号自动避让、保存时编码重复防线
- 样品明细产品名称必填（不默认=编码）、单价必须>0
- 明细行📎资料：产品建档后挂产品文件库，未建档提示先保存
- 产品建档写 spec_json 编码参数（serialNo/panelType等落 product 表，明细表参数弃用）
- 询价单产品名称必填、生成编码不再默认填名称
- ProductCodeGenerator 支持 hideGenerate（四选全选自动生成）
- RouterHelper 顶层菜单 path 缺斜杠自动补全（vue-router 防御）
- 用户角色校验放宽、清理销售死表 SQL

+++ jjx-server/src/main/java/com/jjx/sales/service/IQuotationService.java
@@ -42,7 +42,7 @@ public interface IQuotationService {
      * @param quotation 销售报价单
      * @return 结果
      */
-    int insertQuotation(SalesQuotationAddDTO quotation);
+    Long insertQuotation(SalesQuotationAddDTO quotation);
 
     /**
      * 修改销售报价单
